luci-app-airwhu
This commit delete po2lmo source code and change some files to LF.
