--- package/ctcgfw/luci-app-airwhu/files/root/usr/lib/lua/luci/controller/airwhu.lua
@@ -1,5 +1,0 @@
-module("luci.controller.airwhu", package.seeall)
-
-function index()
-        entry({"admin", "services", "AirWHU"}, cbi("airwhu"), _("AirWHU"), 100)
-        end
