+++ package/ctcgfw/luci-app-airwhu/luasrc/controller/airwhu.lua
@@ -1,0 +1,5 @@
+module("luci.controller.airwhu", package.seeall)
+
+function index()
+	entry({"admin", "services", "AirWHU"}, cbi("airwhu"), _("AirWHU"), 100)
+	end
