+++ package/ctcgfw/luci-app-airwhu/root/etc/config/airwhu
@@ -3,6 +3,7 @@ config switch
 	option enableauth
 	option startwithboot
 	option enableipv6
+
 config auth
 	option Username
 	option Password
