--- package/ctcgfw/luci-app-airwhu/tools/po2lmo/Makefile
@@ -1,12 +1,0 @@
-
-INSTALL = install
-PREFIX  = /usr/bin
-
-po2lmo: src/po2lmo.o src/template_lmo.o
-	$(CC) $(LDFLAGS) -o src/po2lmo src/po2lmo.o src/template_lmo.o
-
-install:
-	$(INSTALL) -m 755 src/po2lmo $(PREFIX)
-
-clean:
-	$(RM) src/po2lmo src/*.o
